Create inital modal for the detailed groupup view
